+++ src/apps/NODAL_ACCUMULATION_3D.cpp
@@ -76,8 +76,8 @@ NODAL_ACCUMULATION_3D::~NODAL_ACCUMULATION_3D()
 
 void NODAL_ACCUMULATION_3D::setUp(VariantID vid, size_t RAJAPERF_UNUSED_ARG(tune_idx))
 {
-  allocAndInitDataConst(m_x, m_nodal_array_length, 0.0, vid);
-  allocAndInitDataConst(m_vol, m_zonal_array_length, 1.0, vid);
+  allocAndInitDataConst(m_x, m_nodal_array_length, (float) 0.0, vid);
+  allocAndInitDataConst(m_vol, m_zonal_array_length, (float) 1.0, vid);
 }
 
 void NODAL_ACCUMULATION_3D::updateChecksum(VariantID vid, size_t tune_idx)
