+++ src/apps/VOL3D.cpp
@@ -73,16 +73,16 @@ VOL3D::~VOL3D()
 
 void VOL3D::setUp(VariantID vid, size_t RAJAPERF_UNUSED_ARG(tune_idx))
 {
-  allocAndInitDataConst(m_x, m_array_length, 0.0, vid);
-  allocAndInitDataConst(m_y, m_array_length, 0.0, vid);
-  allocAndInitDataConst(m_z, m_array_length, 0.0, vid);
+  allocAndInitDataConst(m_x, m_array_length, (float) 0.0, vid);
+  allocAndInitDataConst(m_y, m_array_length, (float) 0.0, vid);
+  allocAndInitDataConst(m_z, m_array_length, (float) 0.0, vid);
 
   Real_type dx = 0.3;
   Real_type dy = 0.2;
   Real_type dz = 0.1;
   setMeshPositions_3d(m_x, dx, m_y, dy, m_z, dz, *m_domain);
 
-  allocAndInitDataConst(m_vol, m_array_length, 0.0, vid);
+  allocAndInitDataConst(m_vol, m_array_length, (float) 0.0, vid);
 
   m_vnormq = 0.083333333333333333; /* vnormq = 1/12 */
 }
