+++ src/basic/PI_ATOMIC-OMP.cpp
@@ -83,7 +83,7 @@ void PI_ATOMIC::runOpenMPVariant(VariantID vid, size_t RAJAPERF_UNUSED_ARG(tune_
         RAJA::forall<RAJA::omp_parallel_for_exec>(
           RAJA::RangeSegment(ibegin, iend), [=](Index_type i) {
             double x = (double(i) + 0.5) * dx;
-            RAJA::atomicAdd<RAJA::omp_atomic>(pi, dx / (1.0 + x * x));
+            RAJA::atomicAdd<RAJA::omp_atomic>(pi, (float) (dx / (1.0 + x * x)));
         });
         *pi *= 4.0;
